ARM64 LPC: add LPC driver for Hip06
This patch introduce a driver to support low-pin-count(LPC) interface for
Hip06 platform. Based on this driver, some LPC cycles can be built to transfer
the I/O port and relevant data between the host device and the peripherals
attached on the LPC. In this way, the host can control the periperals, such as
UART ipmi-bt.

This patch is based on the indirect-IO introduced by last patch. After this
patch is applied, ipmi-bt interface works well without any change.

Signed-off-by: zhichang.yuan <yuanzhichang@hisilicon.com>