ARM64 LPC: LPC driver implementation on Hip06
On Hip06, the accesses to LPC peripherals work in an indirect way. A
corresponding LPC driver need to configure some registers in LPC master
direclty, then the real accesses on LPC slave devices were finished by the
master controller.
In this driver, a LPC specific in/out pair is implemented to read/write data
to the LPC slave peripherals. This device specific in/out pair will be called
by the redefined generic in/out after successful register to the indirectIO.

This patch implement the relevant driver for Hip06 LPC.

Signed-off-by: zhichang.yuan <yuanzhichang@hisilicon.com>