ARM64 LPC: Indirect ISA port IO introduced
On some platforms, there are some legacy ISA I/O devices connected to
low-pin-count interface. When accessing thos legacy devices, some
device-specific port addresses are used to target the corresponding devices.
For example, 0x2f8 is for UART, 0xe4 is for ipmi-bt. This way is different
from the x86 direct IO as I/O space doesn't exist and the MMIO in using on
many architectures.
To drive these devices, this patch introduces a method named indirect-IO.
In this method, the {in/out()} will be redefined as out-of-line function. When
upper layer drivers call {in/out()} with those known legacy port addresses to
access the peripherals, the processing will be redireted to hooking functions
registered specific for the target peripherals. Through this way, those upper
layer drivers which depend on {in/out()} can work well without any changes,
such as ipmi system interface driver.

Signed-off-by: zhichang.yuan <yuanzhichang@hisilicon.com>
Signed-off-by: Gabriele Paoloni <gabriele.paoloni@huawei.com>