ARM64 LPC: Indirect ISA port IO introduced
For arm64, there is no I/O space existing in other architectural platforms,
such as X86. Most I/O accesses there are achieved based on MMIO. But for some
arm64 SoCs, such as Hip06, when accessing the legacy ISA devices connected to
LPC, those known I/O ports are used to locate the corresponding target
devices, for example, 0x2f8 is for UART, 0xe4 is for ipmi-bt. It is different
from the normal MMIO mode in using where only CPU addresses are visible.

To drive these devices, this patch introduces a method named indirect-IO.
The basic ideas of this mechanism are:
1) setup a mapping between LPC bus I/O and linux PIO, then kernel can call
in/out() accessors with linux PIO to perform the I/O wanted.
2) redirect the in/out() to the bus-relevant hooking functions by which I/O
operations can be performed.
Through this way, those upper layer drivers drivers which depend on in/out can
run on Hip06 without any changes.

This patch extend the existing pci_register_io_range() helper function to be
suitable for indirect-IO, rather than only for PCI.

Signed-off-by: zhichang.yuan <yuanzhichang@hisilicon.com>
Signed-off-by: Gabriele Paoloni <gabriele.paoloni@huawei.com>