ARM64 LPC: LPC driver implementation on Hip06
On Hip06, the accesses to LPC peripherals work in an indirect-IO way as MMIO
is not supported there.

A corresponding LPC driver setup the IO space mapping between bus I/O and
linux PIO, and register the bus-specific in/out() hook to perform the I/O
operations on the connected peripherals.

This patch implement the relevant DTS and ACPI driver for Hip06 LPC.
Cooperating with indirect-IO, ipmi messages is in service without any changes
on ipmi driver.

Signed-off-by: zhichang.yuan <yuanzhichang@hisilicon.com>
Signed-off-by: Gabriele Paoloni <gabriele.paoloni@huawei.com>